+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Added caching logic for Gsym resolvers to `symbolize::Symbolizer`
+
+
 0.2.0-alpha.8
 -------------
 - Fixed build failure when `dwarf` feature is not enabled